requirements.txt
torch>=2.0.0
pandas>=2.0.0
numpy>=1.23.0
tqdm>=4.65.0
pyro-ppl>=1.9.0
pytest>=7.3.1
jupyter>=1.0.0
matplotlib>=3.7.1
seaborn>=0.12.2
networkx>=3.0
scikit-learn>=1.0.0
pymsm>=0.1.0